- Friend list added - Match against friend added

+++ app/src/main/java/com/example/mlallemant/mentalbattle/UI/Lobby/PlayAsGuest.java
@@ -2,9 +2,7 @@ package com.example.mlallemant.mentalbattle.UI.Lobby;
 
 import android.content.Context;
 import android.content.Intent;
-import android.os.AsyncTask;
 import android.os.Bundle;
-import android.os.Handler;
 import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
@@ -18,10 +16,11 @@ import android.widget.Toast;
 
 import com.example.mlallemant.mentalbattle.R;
 import com.example.mlallemant.mentalbattle.UI.Game.GameActivity;
-import com.example.mlallemant.mentalbattle.UI.LoginActivity;
+import com.example.mlallemant.mentalbattle.UI.Login.LoginActivity;
 import com.example.mlallemant.mentalbattle.Utils.DatabaseManager;
 import com.example.mlallemant.mentalbattle.Utils.Game;
 import com.example.mlallemant.mentalbattle.Utils.Player;
+import com.example.mlallemant.mentalbattle.Utils.SearchGameTask;
 import com.example.mlallemant.mentalbattle.Utils.Utils;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
@@ -48,7 +47,7 @@ public class PlayAsGuest extends AppCompatActivity {
     private FirebaseAuth mAuth;
     private Player currentPlayer;
     private Game currentGame;
-    private SearchGameTask searchGameTask;
+    public SearchGameTask searchGameTask;
     private DatabaseManager db;
 
     private Boolean appGoesToBackground = false;
@@ -70,10 +69,11 @@ public class PlayAsGuest extends AppCompatActivity {
         super.onStop();
 
         if (!appGoesToBackground) {
-            db.deletePlayerSearchingPlayer(currentPlayer);
+            if (currentPlayer != null )  db.deletePlayerSearchingPlayer(currentPlayer);
             if (currentGame != null) db.deleteAvailableGame(currentGame);
             searchGameTask.cancel(true);
             searchGameTask = null;
+            mAuth.signOut();
         }
     }
 
@@ -110,7 +110,7 @@ public class PlayAsGuest extends AppCompatActivity {
                 db.deletePlayerSearchingPlayer(currentPlayer);
                 if (currentGame != null) db.deleteAvailableGame(currentGame);
                 mAuth.signOut();
-                finish();
+                launchLoginActivity();
             }
         });
 
@@ -160,120 +160,51 @@ public class PlayAsGuest extends AppCompatActivity {
     }
 
 
-    public void launchSearchingGameTask(){
-        Handler handler = new Handler();
-        Runnable r = new Runnable() {
+    private void launchSearchingGameTask(){
+        searchGameTask = new SearchGameTask(new SearchGameTask.AsyncResponse() {
             @Override
-            public void run() {
-               searchGameTask = new SearchGameTask();
-               searchGameTask.execute("");
-            }
-        };
-        handler.postDelayed(r, 100);
+            public void onFinishTask(Game game) {
+                updateUI(game);}
+        });
+        searchGameTask.setParams(currentPlayer, currentGame);
+        searchGameTask.execute("");
     }
 
 
-    private class SearchGameTask extends AsyncTask<String, Void, Game> {
+    private void updateUI(Game game){
+        currentGame = searchGameTask.getCurrentGame();
+        currentPlayer = searchGameTask.getCurrentPlayer();
 
-        @Override
-        protected Game doInBackground(String... urls) {
-            Game returnGame = null;
+        if (game != null) {
+            if ((!game.getPlayer1().getId().equals("")) && (!game.getPlayer2().getId().equals(""))) {
+                String text = "Player found !";
+                tv_log.setText(text);
 
-            try{
-                long start_time = System.currentTimeMillis();
-                long wait_time = Utils.SEARCH_TIME;
-                long end_time = start_time + wait_time;
+                //We insert the currentGame in inProgress
+                db.insertInProgressGame(game);
+                //We get the current, so we can listen only this game now
+                db.initListenerCurrentGame(game);
+                //And also delete the availableGame
+                db.deleteAvailableGame(game);
 
-                Game availableGame = db.findAvailableGame();
+                //Launch game
+                Intent intent = new Intent(PlayAsGuest.this, GameActivity.class);
+                intent.putExtra("idGame", game.getId());
+                intent.putExtra("currentPlayerId", currentPlayer.getId());
+                startActivity(intent);
+                finish();
 
-                if(availableGame == null){
-                    // if no game available, we create one
-                    Log.e(TAG, "no game available");
-                    Player tmpPlayer = new Player("","",0);
-                    String id = getRandomId();
-                    Game game = new Game(id, currentPlayer, tmpPlayer);
-                    db.insertAvailableGame(game);
-                    currentGame = game;
-
-                    while(System.currentTimeMillis() < end_time){
-                        Game tmpGame = db.getAvailableGame(game);
-
-                        if (tmpGame != null && tmpGame.getPlayer1() != null && tmpGame.getPlayer2() != null) {
-                            if ((!tmpGame.getPlayer1().getId().equals("")) && (!tmpGame.getPlayer2().getId().equals(""))) {
-                                returnGame = tmpGame;
-                                break;
-                            }
-                        }
-                        returnGame = tmpGame;
-                    }
-                } else {
-                    // else we insert player1 in game available
-                    Log.e(TAG, "Game available");
-                    if (availableGame.getPlayer1().getId().equals("")){
-                        db.insertPlayer1InAvailableGame(currentPlayer, availableGame);
-                        Log.e(TAG, "Game Player1 inserted");
-                    } else {
-                        db.insertPlayer2InAvailableGame(currentPlayer, availableGame);
-                        Log.e(TAG, "Game Player2 inserted");
-                    }
-
-                    while(true){
-                        returnGame = db.getAvailableGame(availableGame);
-                        if (returnGame != null && returnGame.getPlayer1() != null && returnGame.getPlayer2() != null) {
-                            if ((!returnGame.getPlayer1().getId().equals("")) && (!returnGame.getPlayer2().getId().equals(""))) {
-                                break;
-                            }
-                        }
-                    }
-                    currentGame = returnGame;
-                }
-
-            }catch (Exception e){
-                e.printStackTrace();
-            }
-
-            return returnGame;
-        }
-
-        @Override
-        protected void onPostExecute(Game game) {
-
-            if (game != null) {
-                Log.e(TAG, game.getId());
-                Log.e(TAG, game.getPlayer1().getId());
-                Log.e(TAG, game.getPlayer2().getId());
-                if ((!game.getPlayer1().getId().equals("")) && (!game.getPlayer2().getId().equals(""))) {
-                    String text = "Player found !";
-                    tv_log.setText(text);
-
-                    //We insert the currentGame in inProgress
-                    db.insertInProgressGame(game);
-                    //We get the current, so we can listen only this game now
-                    db.initListenerCurrentGame(game);
-                    //And also delete the availableGame
-                    db.deleteAvailableGame(game);
-
-                    //Launch game
-                    Intent intent = new Intent(PlayAsGuest.this, GameActivity.class);
-                    intent.putExtra("idGame", game.getId());
-                    intent.putExtra("currentPlayerId", currentPlayer.getId());
-                    startActivity(intent);
-                    finish();
-
-                } else {
-                    db.deleteAvailableGame(game);
-                    db.deletePlayerSearchingPlayer(currentPlayer);
-                    String text = "No player found, try again";
-                    tv_log.setText(text);
-                    et_guest_name.setEnabled(true);
-                    btn_play.setEnabled(true);
-                }
-                pb_btn_play.setVisibility(View.GONE);
             } else {
-                searchGameTask = new SearchGameTask();
-                searchGameTask.execute();
+                db.deleteAvailableGame(game);
+                db.deletePlayerSearchingPlayer(currentPlayer);
+                String text = "No player found, try again";
+                tv_log.setText(text);
+                et_guest_name.setEnabled(true);
+                btn_play.setEnabled(true);
             }
-
+            pb_btn_play.setVisibility(View.GONE);
+        } else {
+            launchSearchingGameTask();
         }
     }
 
@@ -294,6 +225,12 @@ public class PlayAsGuest extends AppCompatActivity {
 
     private static String getRandomId(){
         return UUID.randomUUID().toString();
+    }
+
+    private void launchLoginActivity(){
+        Intent intent = new Intent(PlayAsGuest.this, LoginActivity.class);
+        startActivity(intent);
+        finish();
     }
 
 }
